Visio Nocturna
I made a game where you start out in a lucid dream, but it's a nightmare. The player finds themself in front of a door and they have to find five keys to escape. If they don't, there is a jump scare, they lose, and the next time they play, the creature that's after them remembers they died before and makes them slower.